rewrite the 2nd-init detection algorithm
this update to 2nd-init makes it auto-detect if the heap space is
executable and if it is uses it, if it isn't attempts to inject into
the code segment.

this allows the same 2nd-init code to run on froyo and gingerbread
where froyo had executable heap space and gingerbread did not for
certain phone models.